requirements.txt
streamlit==1.25.0
requests==2.31.0
Pillow==9.5.0

# requests-oauthlib==1.3.1
# base64
# io
# request